feat: 포스터 삭제 API 로직 변경 #5 Poster - deletePoster: 이미지 URL이 존재할 때만 S3 내 이미지 삭제 로직이 동작하도록 변경

+++ src/main/java/fete/be/domain/poster/persistence/Poster.java
@@ -137,7 +137,9 @@ public class Poster {
 
         // S3에서 이미지 삭제
         for (PosterImage posterImage : poster.posterImages) {
-            imageUploadService.deleteFile(posterImage.getImageUrl());
+            if (!posterImage.getImageUrl().isBlank()) {
+                imageUploadService.deleteFile(posterImage.getImageUrl());
+            }
         }
         // DB에서 이미지 삭제
         poster.posterImages.clear();
